requirements.txt
pyTelegramBotAPI==4.15.4
google-genai==0.3.0
flask==3.0.2
werkzeug==3.0.1
requests==2.31.0
python-dotenv
python-i18n==0.3.9
cryptography==42.0.8
